Move button_to helper into core_ext/rails
The `button_to` helper is currently implemented as an override of the `button_to`
method provided by ActionView, and therefore should only be available in Rails.
