+++ spec/helpers/button_to_helper_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-require 'bh/helpers/button_to_helper'
+require 'bh/core_ext/rails/button_to_helper'
 require 'bh/helpers/navbar_helper'
 
 include Bh::ButtonToHelper
